Add sad path testing for removing favorite endpoint

+++ app/controllers/api/v1/favorites_controller.rb
@@ -20,7 +20,7 @@ class Api::V1::FavoritesController < ApplicationController
   end
 
   def destroy
-    if user_by_api_key
+    if user_by_api_key && user_favorite?(clean_location)
       facade = FavoritesFacade.new(user_by_api_key)
       facade.delete_favorite(clean_location)
       render json:'Accepted',status:202
@@ -32,7 +32,11 @@ class Api::V1::FavoritesController < ApplicationController
   private
 
   def user_by_api_key
-    user = User.find_by(api_key: favorite_params["api_key"])
+    User.find_by(api_key: favorite_params["api_key"])
+  end
+
+  def user_favorite?(location)
+    user_by_api_key.favorites.pluck(:location).include?(location)
   end
 
   def clean_location
